% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gates_blastn}
\alias{gates_blastn}
\title{gates_blastn}
\format{\preformatted{'data.frame':	1689 obs. of  12 variables:
 $ Query_id        : int  316 316 316 316 316 316 316 316 316 316 ...
 $ Subject_id      : Factor w/ 1808 levels "A16379.1.1485",..: 410 411 306 66 409 1556 239 1301 1255 1034 ...
 $ percent_identity: num  98.1 98.1 98.1 98.1 97.7 ...
 $ alignment_length: int  214 214 214 214 214 152 152 149 149 149 ...
 $ mismatches      : int  4 4 4 4 5 1 1 1 1 1 ...
 $ gap_openings    : int  0 0 0 0 0 0 0 0 0 0 ...
 $ q_start         : int  3 3 3 3 3 3 3 3 3 3 ...
 $ q_end           : int  216 216 216 216 216 154 154 151 151 151 ...
 $ s_start         : int  356 356 356 356 356 362 354 336 336 364 ...
 $ s_end           : int  143 143 143 143 143 211 203 188 188 216 ...
 $ e_value         : num  4e-109 4e-109 4e-109 4e-109 1e-106 ...
 $ bit_score       : num  392 392 392 392 385 293 293 287 287 287 ...
}}
\usage{
gates_blastn
}
\description{
Blast results from Gates 1000 dataset.
}
\examples{
gates_blastn
}
\keyword{datasets}

