% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gates_tax}
\alias{gates_tax}
\title{gates_tax}
\format{A data frame with columns:
\describe{
 \item{OTU}{}
 \item{Kingdom}{}
 \item{Phylum}{}
 \item{Class}{}
 \item{Order}{}
 \item{Family}{}
 \item{Genus}{}
 \item{Species}{}
}}
\usage{
gates_tax
}
\description{
Taxonomy data for the Gates 1000 dataset according to NCBI.
}
\examples{
gates_tax
}
\keyword{datasets}

