% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_blast_m9}
\alias{read_blast_m9}
\title{Convert a blast result obtained while using the -m 9 flag with blastall
into a data frame.}
\usage{
read_blast_m9(path)
}
\arguments{
\item{path}{The path to the fasta file.}
}
\description{
Convert a blast result obtained while using the -m 9 flag with blastall
into a data frame.
}
\examples{
\dontrun{

read_blast_m9("path/to/file.blastn")

}
}

