% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seq_length_tbl}
\alias{seq_length_tbl}
\title{Convert a fasta file into a data frame containing the decription of each
sequence in the first column and the length of that sequence in the second
column.}
\usage{
seq_length_tbl(path, description = "^>", comment = "^;")
}
\arguments{
\item{path}{The path to the fasta file.}

\item{description}{A regular expression indicating which lines contain sequence descriptions.}

\item{comment}{A regular expression indicating which lines contain comments.}
}
\description{
Convert a fasta file into a data frame containing the decription of each
sequence in the first column and the length of that sequence in the second
column.
}
\examples{
\dontrun{

# Normal use
seq_length_tbl("path/to/file.fasta")

# Comments are lines where number sign is first character
seq_length_tbl("path/to/file.fasta", comment = "^#")

}
}

