% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{silva_NR_typestrains}
\alias{silva_NR_typestrains}
\title{silva_NR_typestrains.}
\format{A data frame with columns:
\describe{
 \item{Accession_Number}{The accession number of the organism.}
 \item{Sequence}{The nucleotide sequence.}
}}
\source{
\url{http://www.arb-silva.de/}, downloaded 2014-10-10
}
\usage{
silva_NR_typestrains
}
\description{
The Silva non-redundant typestrain database for the small
ribosomal subunit, release r119.
}
\examples{
silva_NR_typestrains
}
\keyword{datasets}

