% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{silva_typestrainNR_taxonomy}
\alias{silva_typestrainNR_taxonomy}
\title{silva_typestrainNR_taxonomy}
\format{A data frame with columns:
\describe{
 \item{Accession_Number}{The accession number of the organism.}
 \item{Kingdom}{}
 \item{Phylum}{}
 \item{Class}{}
 \item{Order}{}
 \item{Family}{}
 \item{Genus}{}
 \item{Species}{}
}}
\source{
\url{http://www.arb-silva.de/}, downloaded 2014-10-10
}
\usage{
silva_typestrainNR_taxonomy
}
\description{
Taxonomy data for the Silva non-redundant typestrain database for the small
ribosomal subunit, release r119.
}
\examples{
silva_typestrainNR_taxonomy
}
\keyword{datasets}

