% Generated by roxygen2 (4.0.2): do not edit by hand
\name{warp_tbl}
\alias{warp_tbl}
\title{Given a data frame and two column names, this function returns a data frame
where groupings of the first variable contain differences with respect to the
second variable.}
\usage{
warp_tbl(tbl, grouping, warp_by)
}
\arguments{
\item{tbl}{A data frame with at least two columns}

\item{grouping}{The column name specifying how to group the rows.}

\item{warp_by}{The column name specifying the differences within groups to look for.}
}
\description{
Given a data frame and two column names, this function returns a data frame
where groupings of the first variable contain differences with respect to the
second variable.
}
\examples{
\dontrun{

fruit_prices <- data.frame(Fruit = c("Apple", "Apple", "Orange", "Orange", "Pear", "Pear"),
                           Price = c(1.2, 1.3, 1.1, 1.1, 1.4, 1.1))
warp_tbl(fruit_prices, "Fruit", "Price")
warp_tbl(fruit_prices, "Price", "Fruit")

}
}

