\name{warppipe-package}
\alias{warppipe-package}
\alias{warppipe}
\docType{package}
\title{
Phylogenetic Annotation Pipeine for Biological Sequences
}
\description{
This package provides tools for the analysis of biological
sequences in relation to their phylogenetics.
}
\author{
Sean Kross
}
\seealso{
\itemize{
\item \code{\link{seq_tbl}}
\item \code{\link{seq_length_tbl}}
\item \code{\link{read_blast_m9}}
}
}
