% Generated by roxygen2 (4.0.2): do not edit by hand
\name{warp_tbl}
\alias{warp_tbl}
\title{Filter out promising blast hits given data frames created from other
functions in this package.}
\usage{
warp_tbl(blast, tbl, n = 1, ...)
}
\arguments{
\item{blast}{A data frame converted from a blast result by using the function
read_blast_m9().}

\item{tbl}{A data frame of a fasta file converted using seq_length_tbl().}

\item{n}{The top n hits to be returned.}

\item{...}{Logical conditions to further filter the table. Combine conditions
with \code{&}.}
}
\description{
Filter out promising blast hits given data frames created from other
functions in this package.
}
\examples{
\dontrun{

# Setup
blast_result <- read_blast_m9("result.blastn")
sequence_table <- seq_length_tbl("db.fasta")

# Get top hit for each query
warp_tbl(blast_result, sequence_table)

# Get top 5 hits for each query
warp_tbl(blast_result, sequence_table, n = 5)

# Get top 3 hits for each query where bit_score is greater than 400.
warp_tbl(blast_result, sequence_table, n = 3, bit_score > 400)

}
}

